DOT: Move package and handle errors
